requirements.txt
PyQt5>=5.6
stegano~=0.9.8
pycryptodome